net: phy: fix 88E1510 PHY initialization
The initialization of the 88e1510 was wrong in that
it performed the 88e1518 configuration which then calls
the 88e1111s initialization.

The 88e1111s initialization is incompatible with this phy, especially when
it comes to configuring RGMII.  Also, the 88e151X phys do not support
RTBI mode.

This code changes the initialization of the 88e1510 to use the proper
registers for RGMII.  Note that I have not been able to test the different
timing modes with my board so I am only using PHY_INTERFACE_MODE_RGMII.

Also, it may be that instead of making the change here the 88e1518
configuration function should be changed.

Change-Id: I0687d94563bd0c8038d0abae04735dcae249bb7d
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>