Fix a number of issues with eMMC and adds some enhancements.
There are a number of changes with this checkin.  One problem with the old
code is it would crash if a device were not plugged in and would also
attempt to remove slots where no device is connected.  This code no longer
removes slots when nothing is plugged in.

getting *slot and *host now uses cavium_get_slot(mmc) and
cavium_get_host(mmc) respectively.

The initialization sequence has also been cleaned up and now more closely
follows the generic mmc driver init sequence.

When powering off devices with a shared power GPIO line it first checks to
make sure that all shared devices are being powered down before actually
turning it off.  This requires that the power-gpios parameter be placed
in the mmc_1_4 section and not in the slot section.

Short timeouts have been extended although this likely isn't necessary.

If mmc_initialize() is called more than once, make sure to uninitialize
everything first.

The driver name is now in a #define

Fixed the initialization of host->dev_index and do not use host->node.

Signed-off-by: Aaron Williams <aaron.williams@cavium.com>