phy: marvell: cp110: add basic AP (auto-negotiation) support
During AP the proper mode is auto-negotiated and the mac, pcs and serdes
configuration are done by the firmware loaded to the MG's CM3 for appropriate
negotiated mode. Therefore there is no need to configure the mac, pcs and
serdes from u-boot. The only thing that need to be setup is powering up
the comphy, which is done through Common PHY<n> Configuration 1 Register
(CP0: 0xF2441000, CP1: 0xF4441000). This step can't be done by MG's CM3,
since it doesn't have an access to this register-set (but it has access to
the network registers like: MG, AP, MAC, PCS, Serdes etc.)

Change-Id: I8498a2a452d95c8993c483272556838b902dabf1
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-by: Igal Liberman <igall@marvell.com>