net: switch: mv88e6xxx: add port setup in initialization
With default setting in port control register, switch works well; but
linux kernel will change the port control setting, for example, it will
add tag for egress frames in cpu port; then after reboot in kernel,
the changed port control setting is kept in u-boot, but the changed
setting does not work for switch in u-boot.
This patch adds port setup in initialization to make sure that switch
can work properly in u-boot.

Change-Id: Ic380f80f712707003668172d6092d1603f54a570
Signed-off-by: Ken Ma <make@marvell.com>
Signed-off-by: Igal Liberman <igall@marvell.com>