efuse: mvebu: add DM and dts support for efuse driver
Efuse feature is fixed with hardcoded for a38x efuse, which is not
flexible for different mvebu SoC chips.
This patch adds driver model and dts for efuse driver and efuse
dts node in dtsi file and enables it on dts file.
Besides, it removes those older mvebu fuse codes.

Change-Id: I4591f8356f9ed0e82e36c0a20911d66a4cb1d486
Signed-off-by: zachary zhang <zhangzg@marvell.com>