ehci: mvebu: toggle vbus for EHCI start and stop
Currently, the VBUS for EHCI is fixed on 5v and never toggle it
when do usb start and usb stop.
According the USB2.0 SPEC, the timing between VBUS on and USB
enumeration should be 200ms. But now the timing depends on
when to run the command "usb start" and it maybe a long time. It
is not follow the USB2.0 SPEC and will cause some u-disk misdetection.
This patch add toggle VBUS for EHCI to follow the time requirement
in USB2.0 SPEC and also add marvell.ehci-usb document.

Change-Id: I5b10b6a391ef1cb13df627fb9eb3126509862d9f
Signed-off-by: allen yan <yanwei@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>