BGX: Adjust lane_to_sds based on BGX-ENABLE.* macro
On SFF8104 BGX0 is configured as QSGMII or XFI interface based on the GPIO settings.
When BGX0 is in XFI mode, according to schematics, it uses DLM1 lane1. Recently
BDK has been udpated (v2.02.2) to configure the QLM in XFI_2X1 mode which expects
lmac1 to be configured in XFI mode, this is configured by setting '1' to
BGX-ENABLE-N0.BGXx.Px macro. Adjust qlm-mode to match this so that BDK can trim the
device tree properly.

Parse BGX-ENABLE.N0.BGXx.Px to find which lmac is configured.

These are the macros configured in BDK when BGX0 is configured in XFI mode on
sff8104 board
  BGX-ENABLE.N0.BGX0.P0=0
  BGX-ENABLE.N0.BGX0.P1=1
  BGX-ENABLE.N0.BGX0.P2=0
  BGX-ENABLE.N0.BGX0.P3=0

u-boot still configures BGX0.lmac0, sets lane_to_sds to 3 instead of 2.

Signed-off-by: Chandrakala Chavva <cchavva@caviumnetworks.com>