spi: kirkwood: add direct write mode
This patch ports the linux kernel commit "spi: orion: Add direct
access mode" written by Stefan Roese to u-boot.

This patch adds support for the direct write mode to the kirkwood
SPI driver which is used on the Marvell Armada based SoCs. In this
direct write mode, all data written to a specifically mapped
window (linked to one SPI chip-select on one of the SPI controllers)
will be transferred directly to the SPI bus. Without the need to
control the SPI registers in between. This can improve the SPI
write rate in such cases.

Please note that direct-read and direct-write mode are supported in
different way.
 - Write mode benefits from the SPI direct mode, as the data bytes
   are written head-to-head to the SPI bus, without any additional
   addresses;
 - For direct-read mode, we can read spi data directly from the
   specifically mapped window, for example, the SPI mapped window
   is [0xf9000000, 0xf9ffffff] in armada 7040 DB board, we can read
   spi flahs address n(n < 0x1000000) by reading (0xf9000000 + n).

This mode is described in chapter "43.10 Direct mode" in the Marvell
Armada 70xx/80xx Functional Spec Datasheet.

Change-Id: Ic1e383c982211f511ecf18b55197788f5fe391d7
Signed-off-by: Stefan Roese <sr@denx.de>
Signed-off-by: Ken Ma <make@marvell.com>