arm64: mvebu: a8k: set cpuidle.off=1 in bootargs for A8040 A1/2
Linux kernel cpuidle feature is not supported
in Armada8040 A1/A2 revisions, in order to be able
to enable it by default in Linux kernel, we need
to set in Linux kernel bootargs cpuidle.off=1.

This patch adds $cpuidle variable to set_bootargs.
By default $cpuidle is empty.
For SoC revision is A1/2, set $cpuidle to cpuidle.off=1,
this will disable cpuidle in Linux kernel.

In addition, enable CONFIG_ARCH_MISC_INIT, in order
to use arch_misc_init() for this purpose.

Change-Id: Ie0c98bb0644708e6ab89fa71599e8bf7ea1a59da
Signed-off-by: Igal Liberman <igall@marvell.com>