mtd: rawnand: pxa3xx: add SoC-specific enable function
Add SoC-dependent function to the NAND driver for selecting
NAND interface in DEVBUS MUX register.
This selection is done in the BootROM if the boot device is NAND,
but may be missing othervise.
The NAND is selected only if it is enabled in the DT file.
This patch is fixing NAND access problems for configurations
that use non-NAND devices for boot (like SPI).

Change-Id: I9cd9b2cbecee0eaa2dd54eacb5ca0a55f4bede3b
Signed-off-by: Igal Liberman <igall@marvell.com>
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>