octeontx: mmc: Rename cavium_mmc to octeontx_mmc and use hw generated regs
This patch renames the cavium_mmc driver to octeontx_mmc.  Additionally it
modifies the driver to use an auto-generated hardware register definition
file in order to make supporting future SOCs easier.

Renaming cavium to octeontx was done due to the recent aquisition of Cavium
by Marvell, Inc.  Note that there are still a few places that need to be
fixed.

A future patch will clean up the changes required for the generic mmc files.

Signed-off-by: Aaron Williams <aaron.williams@cavium.com>