mmc: octeontx: octeontx2: Added octeontx2 support
This patch breaks out setting the bus timing parameters to use a separate
function rather than have duplicate code in different places.  Note that
the meanings for clock and data skew from the device tree change.  In
OcteonTX, these specify values that go into the sample register.  For
OcteonTX2, however, these specify the delay in picoseconds due to the
improved method in which the sample delays are set.

Signed-off-by: Aaron Williams <aaron.williams@cavium.com>