octeontx: mtd: Add mtd partitioning support
This patch enables MTD partitioning support.  A default partition table
has also been added which defines 6MB for the boot partition, 9MB for
data, and 64K for the environment, since 64K is the minimum supported
size.  I do not know what is after the environment.

Note that it is possible for U-Boot to add the partitioning to the
device tree so it can be passed on to Linux but this is not enabled
at this time.

Signed-off-by: Aaron Williams <aaron.williams@cavium.com>