octeontx: nand: fix the PCI class id
This fixes the PCI class ID so that the OcteonTX NAND driver is
properly probed.  Additionally, this adds functionality to
shut down the devices at the end.

Note: for some unknown reason when the NAND driver is used it
causes the Linux kernel to crash when asynchronous interrupts
are enabled.

Signed-off-by: Aaron Williams <aaron.williams@cavium.com>