mmc: octeontx/octeontx2: Fix obtaining bus width from device tree
The OcteonTX driver was using the old obsolete parameter
"cavium,bus-max-width" instead of "bus-width" in the device tree to
set the maximum width of the bus.  This causes devices to fail when
the physical bus is smaller than what the device supports.

This patch also cleans up some minor white space issues.

Change-Id: I512f1b59ca4a3a61013e9847e510b7713bf44d53
Signed-off-by: Aaron Williams <aaron.williams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/1478
Tested-by: sa_ip-sw-jenkins
Tested-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>
Reviewed-by: Kostya Porotchkin <Kostya.Porotchkin@cavium.com>
Reviewed-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>