env: sf: add support for fio spi1 env with fdt string
Use fdt string "u-boot,env" in spi flash nodes, nodes with
compatible "spi-flash", to make user requested device
as storage medium for u-boot environment.
With 9xx chips, spi controllers count is 2 and each can support
two flash devices on each chip select, so it makes sense to
push this logic to internal sdk for easy custom env device.
If not defined, default values for controller and chip-select
will be 0.

This may not be upstreamable as it is required to have build
option for each board so that it is fixed at compile time.

Change-Id: Ib6d733c9636800f788bc7140dd92371bab617c4a
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/3923
Tested-by: sa_ip-sw-jenkins
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>