Fix MMC support in ASIM and add emulator support
This patch adds several more areas where timing operations
are skipped when running inside ASIM.  In addition, the
pre-init command was removed since it is not needed and
Linux does not use it nor the core U-Boot MMC.

Emulator support was also added.  For now, the emulator
support is identical to ASIM, however that can change
in the future.

Change-Id: I59a87be12ae6c1661b3fec2ec8411531a183a4be
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/8731
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Roy Franz <Roy.Franz@cavium.com>