mmc: octeontx: Enable I/O power control
This patch allows the OcteonTX driver to automatically control
the I/O power so it can switch between 3.3 and 1.8v devices.
This is required when both SD and eMMC devices share the same
bus so that HS400 mode can be supported.  When a 1.8v device
such as eMMC is selected, all of the other I/O supplies are
disabled.  Similarly, when a 3.3v device like an SD card is
selected the 1.8v I/O is disabled.

This has been tested with the EBB9604, EBB9304 and EBB9504
boards.  This eliminates the need to issue the
gpio set gpio@6c_307, etc. as well as editing the device
tree to select which devices are available.

Change-Id: I9df96babc0a73461a31006583af0917587f97ee8
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/10560
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>