pinctrl: armada-37xx: remove SPI CS control from UART2 group
(This code is not "correct" for upstreaming, but works for our needs)

Whenever we touch any GPIO in the UART2 group, there is a request
to switch the entire UART2 pinmux group to GPIO mode.  This affects
the SPI_CS2 and SPI_CS3 and effectively steals them from the SPI
driver.

Remove the SPI CS control bits from the UART2 group.  If we were trying
to access a SPI CS as a GPIO, the associated SPI group will do the right
thing.  This code is likely broken for UART2 as UART function, because
the SPI CSs are not disabled correctly.

Change-Id: I53ad6b590a921ac33098481c6960e2b89b8e869c
Signed-off-by: Patrick Williams <alpawi@amazon.com>