octeontx: serial: Added pci-bootcmd input driver
This adds a pseudo serial driver and console driver
using shared memory accessed across the PCIe bus.
This allows the host utility cavm-remote-bootcmd
and cavm-remote-load to send commands directly to
the U-Boot command line.

In order to enable this, iomux support is also
enabled.  The stdin environment variable must be
set to serial,pci-bootcmd to enable this input.

Change-Id: I4168d20a693d039ae354910b071316c415b9e26e
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/11573
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>