usb: xhci: octeontx2: Fix RX issue with short packets
There appears to be an issue in OcteonTX2 where
if the receive buffer crosses a 64K boundary
and partial data is received that fits in
the first part of the buffer in front of
the boundary then the OcteonTX2 XHCI engine
will generate two event responses.  The first
response will have a completion code of a
short packet, which is correct, however the
length field is wrong.

In addition, a second event is also generated
with a success completion code with another
invalid length.

Due to this corruption, receive buffers where
short packets can occur must never cross a 64K
boundary with the OcteonTX2 SoCs.

This patch adds a 64K bounce buffer aligned to
64K.  In the event that an input pipe is used,
the bounce buffer will be used instead of the
user-supplied buffer and the data will be copied
from the bounce buffer to the user buffer.

Change-Id: I5cbbf28ae14538c1a970023c1daee2695b20978e
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/12524
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Aaron Williams <Aaron.Williams@cavium.com>