mmc: octeontx2: Fix numerous issues
This patch removes the work-around code for resetting
the eMMC subsystem when adjusting the timing since this
is no longer needed.

The eMMC tuning code has been rewritten to match the code in the
Linux kernel driver.  Instead of starting at the end and choosing
the second good value, it goes through all values and chooses the
center of the longest run of good values.

Calculating the delays for the taps now rounds up so timing
values will not be too short.

Switch to unused busid 3 when performing calibration to prevent
any chances of causing glitches for the current mmc bus as well
as reset the subsystem.

Increase DMA timeouts for read and write operations.  The old
settings were causing premature timeouts with some devices.

Switch to mmc bus 3 when changing the I/O supply voltage to
eliminate any chances of glitches due to switching.

Removed timing glitch code since the current CNF95XX pass "A0"
no longer exhibits this issue.

Change-Id: Id594e37ad88c062b15263ce114dfd6bc4a99a95a
Signed-off-by: Aaron Williams <aaron.williams@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/13141
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>