mmc: octeontx: Fix HS400 writes and other issues
This fixes the tuning support and writes in HS400 mode.
Part of this involves having two timing values, one
for HS200 and the other for slower modes.  This is
needed because after selecting HS200 mode it briefly
drops down to HS mode before switching back.

The calibration code now clears the start bit after it
is complete.  For some reason this prevents MMC write
operations to generate block CRC errors if it is left
enabled even though the calibration operation has
completed.

When changing the timing register and for some other
operations, switch to bus ID 3 to prevent glitches
with the current bus ID.

This also fixes the response when an error occurs during
DMA so it does not wait to timeout.

This fixes a case where the RCA register is not getting
set for mmc.  It gets initialized right before the
SEND_EXT_CSD command so it's safe to initialize it
at this point.

Change-Id: If116b6156acb571ee9f70d7bac13193eefedd80d
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/14364
Tested-by: Suneel Garapati <sgarapati@caviumnetworks.com>
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>