octeontx: Add PCIe console
This adds code for performing console read/write operations using
a shared memory buffer across the PCIe bus.  This consists of two
drivers.  The first driver is a nexus driver, which handles the
base descriptor for one or more consoles.  The console driver
handles a single individual console.  Note that the code requires
the use of atomics because the nexus data structure can be shared
by multiple entities including ATF.  Atomics are also used
in the console code itself to prevent any corruption from occuring
should a PCIe host console connect during some critical operation.

The functionality is not complete at this point and there is some
likely cleanup in the future.  Currently support for sharing a
console between ATF and U-Boot is missing.  Also, U-Boot should
be able to specify which console to use rather than the first
available console.

Change-Id: I936219827b8835431bd1dc5c55951e03e38ff8ca
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17158
Reviewed-by: Derek Chickles <Derek.Chickles@cavium.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>