octeontx: bootimgup Numerous fixes including preserving ROM scripts
This patch fixes a number of patches.
1. Improve image validation checks by checking more sections and verifying
   any checksums and CRCs if available.
2. Fix code that preserves any ROM scripts to also update the CLIB info blocks
3. Do not remove and probe for the flash pointer, instead use the pointer
   provided by spi_flash_probe_bus_cs().
4. Validate that any ROM scripts being preserved have a valid checksum

Since preserving the ROM scripts involves modifying the data in the
buffer passed on the command line, the data is restored after the
operation completes.  The code is also designed to handle the (likely)
case where the SCP NB1FW INFO ROM script and the SCP TBL1FW INFO ROM
scripts point to the same address.

Change-Id: Ia22f24b24b67fe300c850f9581ff0224f87b0454
Signed-off-by: aaronw <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17579
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Derek Chickles <Derek.Chickles@cavium.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>