octeontx: octeontx2: increase CONFIG_SYS_SCSI_MAX_SCSI_ID to 16
Increase CONFIG_SYS_SCSI_MAX_SCSI_ID to 16 to allow up to 16 ports of
add-in PCIe SATA controllers to work.  This should allow all ports to
work on all common SATA PCIe controllers.
'scsi reset' time on t83 is not affected, and debug prints verify that
the device reported port number (of 1) per controller is used to limit
the ports scanned for the builtin SATA ports.

Change-Id: Ie33858bc63915b3b0db78702154b5ada77a62227
Signed-off-by: Roy Franz <rfranz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/18001
Reviewed-by: Aaron Williams <awilliams@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>