marvell: octeontx/tx2: Enable bootretry env var capability
Enable the ability to set the bootretry environment variable and
have it take effect. By default the time is -1, which means it
is disabled. If enabled, it must be at least 30 seconds so the
user has time to type in a command.

Change-Id: I3ad40c127f2849765b37b83bd8d4172a73cac18f
Signed-off-by: Chickles, Derek <Derek.Chickles@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/18668
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
(cherry picked from commit e61713d0f31928aec198ba824fdc4a98a9129510)
Reviewed-on: https://sj1git1.cavium.com/19361
Tested-by: Suneel Garapati <sgarapati@caviumnetworks.com>
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>