octeontx: pci console: Fix PCI console shutdown
After upgrading to 2019.10 the driver removal process needs to remove
itself from the IO mux by modifying the stdin, stdout and stderr
environment variables.  In addition, we deregister ourselves from
stdin.  Ideally there would be an easier way to register and
deregister ourselves without having to go about modifying
environment variables.

Without this change, the printf after deregistering causes U-Boot
to crash when a printf occurs when it still tries to print to
the now non-existent PCI console driver.

In addition, this updates the Kconfig file to automatically select the
required features.

Change-Id: Ica44222c8a1f9c8a42ad18550214b11cd4f2ed5d
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/20555
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>