mmc: xenon_sdhci: improve voltage switch sequence
Until now in the xenon_sdhci driver the voltage switch
relied on the pad_type selection, which is valid for
the controller found in Armada37xx SoCs, but not in
Armada 7k8k / CN913x. Update the driver to ignore it
when running on the latter.

Also the voltage switch sequence was depending on the
SDHCI_POWER_CONTROL register settings, which is set
and maintained by the generic U-Boot code. It also
refers to the bus voltage (Vcc/Vdd), but not the IO
lanes. In some setups this prevented the actual voltage
switch to be performed, despite the proper connectivity
available. In order to fix that, rely on the
mmc->signal_voltage field passed from the upper layers
to the controller driver.

This patch enables proper operation of the highest
speeds on Armada 7k8k / CN913x SoCs.

Change-Id: I6a5e99a250fdfa444ce7ed112ced3ac6f4abde9d
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/21960
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Grzegorz Jaszczyk <jaszczyk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>