include/configs/mvebu: fix NAND environment location
NAND environment should be aligned to the flash erase size.
For Armada boards with 1GB 4KB page NAND the ersase size
is 256KB.
Additionally the u-boot partition size according to kernel
device tree is 2MB and not 4MB as was stated in the common
config file.
This patch fixes the above sizes.

Change-Id: I8a176d48ceef8b69b03f2cd4277d2f6ec11d013d
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/22747
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>