octeontx2: mmc: Add tuning support for HS400 mode
The eMMC specification does not provide a means to perform tuning in
the HS400 mode and instead relies on using tuning parameters obtained
from HS200 mode.  This is not sufficient.  In many cases values that
are good for HS200 mode do not work for HS400 mode.  We have found
many instances where values right in the middle of a good range
for HS200 fail in HS400 mode on some hardware.

This patch adds support so that after tuning in HS200 mode a block
specified in the device tree is written with some worst-case data
to aid for tuning in HS400 mode if this data is not already present.
After switching to HS400 mode a new tuning operation is performed.
In this case, the driver reads the specified block at every tap
value and chooses the point midway in the longest good run.

This is similar to how it would work had the broken eMMC standard
supported tuning in HS400 mode.

Change-Id: I1c81382d534a179b9dacc58e09e9134bcd15593c
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Signed-off-by: Chandrakala Chavva <Chandrakala.Chavva@cavium.com>
Reviewed-on: https://sj1git1.cavium.com/22715
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>