arm64: mvebu: soc: extend dfx smc handler
Since the dfx register set is going to be marked as secure, more drivers
are going to use smc for accessing some dfx registers. Therefore the dfx
smc routine need to be extended. Additionally macro definitions which
use common dfx smc in a different way are added for further use.

This patch is a preparation for upcoming changes and will prevent mixing
soc and thermal driver changes with upcoming drivers rework.

At the occasion fix naming of first argument passed via smc call which
refers to sub-fid.

Change-Id: I5cf8ac596412c8eecdb13cf7200222c369c636da
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/25050
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>