mmc: octeontx2: Fix hold timing for HS200 and HS400
This patch fixes U-Boot to use calculated tap values for the data and
command output timing rather than the hard-coded tap value of 5.
The hard-coded value usually provided a hold timing that was too
short.  The hold timing should be 800ps for command and data for
HS200 and 800 and 400ps for HS400 mode.

In addition, this patch allows the device tree to override the
hard-coded values.

Change-Id: Iec0654fd31411327386cf974fff6664c9a06633f
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/23545
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>