mmc: octeontx: octeontx2: Fix DMA cleanup operation
It seems that the DMA cleanup code had a mistake in that it was
checking both dma_val and dma_pend when it likely that dma_val will
go to zero before dma_pend.  This also performs retries of the dma
cleanup operation.  A customer reported seeing errors prior to this
change and this patch fixes it.

Change-Id: Ic0a7725d3b9010052b0e10f0b566f413ff701db1
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/25938
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>