Preserve firmware-layout for Linux
This patch changes how octeontx_brd is created.  Previously, several fields
were extracted then cavium,bdk was deleted and a new node was created with
the properties re-added.  The problem was that this removed the
firmware-layout tree.

Instead, all of the properties are read and all except those to be preserved
are deleted.  This leaves firmware-layout in place.  Afterwards, cavium,bdk
is renamed to octeontx_brd.

Change-Id: Ibf41ae2d1fbf7c6b6b06eaef0b36bc729acb3a14
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/32603
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>