arm64: mvebu: force the first spi bus probe early
Current DM discovery skips SPI bus detection when the default
boot/environment media is not located on SPI flash and the
first SPI bus is located on slave CP (for instance CP1/CP2).
In such cases the SPI bus on slave CP controller is not
probed and later probe for the SPI flash on this bus fails.
This patch triggers the first SPI bus detection on early
boot stages.

Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Change-Id: I178e6b537f018a1c9ccac98e387d8d38768fec33
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/47012
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>