drivers: spi: Fix possible issue in xSPI driver
When xSPI will be configured by sf probe, and before any operation
on spi memory SMC call will be executed - it will cause an error, as
xSPI driver will not check internal state of IP core before issuing
STIG command.
Fix issue - verify if xSPI is in STIG mode, and run config again if
it is not.

Change-Id: I41d91de85841a41cd6c3818a2d146022214b3328
Signed-off-by: Witold Sadowski <wsadowski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/47874
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Narinder Dhillon <Narinder.Dhillon@cavium.com>
Reviewed-by: Suneel Garapati <sgarapati@caviumnetworks.com>