efi_loader: Fix coverity warnings for efi tcg2 protocol
Coverity reported 3 warnings on the current code.
CID 331856, 331855, 331854 on the latest scan.

Fix the rest of the warnings by initializing the variables before
passing them to tpm2_get_pcr_info().
In order to avoid future warnings and errors initialize them to 0 within
the function as well, since the values are always OR'ed after querying the
hardware.

Change-Id: I14944b48c3d4761e2a7ab8d231ef38f2fc1b4c86
Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/81983
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>