drivers: pci: Optimize scan for non-ARI devices on Root bridge
Should not scan for device numbers 1-31 as per PCIe spec r3.1, sec 7.3.1
on non-ARI downstream devices.
So add a check if PCI bus bridge is of ROOT PORT type from PCI Express
Capability header and update end variable to skip device numbers 1-31.
Later, during scan if the first device is found to be on ARI chain update
the end variable to max again.

Change-Id: Iaca66c111e433df986f2fecd8f7623d3521b76bf
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/52851
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>