cmd: cn10k SerDes: fix stopTraining() call
This fixes the case, when portm is in multilane mode
and the rx training was started only for the specific
lane (but other than zero) and the training happens to
fail due to the timeout (does not complete before the
timeout). In that case the stopTraining() would be
requested for the wrong lane number.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: I830e2f8836038bf9636110a5e1b3449bf9cdcfaa
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/64907
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>