octeontx: pci console: adjust padding size for RT Linux
Linux and U-Boot must share the same layout of the PCI console structures.
If Linux is built with CONFIG_PREEMPT_RT, the data type of the excl_lock
field (spinlock_t) grows by 44 bytes. That makes the `in_use` field start
at an unaligned offset. When kernel boots and the driver tries to acquire
the console, a data abort exception is raised, crashing the system.

This issue is fixed in Linux by adding 4 bytes padding when PREEMPT_RT
is set and 48 bytes if it is not set. That way every field after the spin
lock object stays at the same 64-bit aligned offset regardless we are
running RT or non-RT kernel, avoiding the data abort and making the whole
struct looks the same in both modes.

Since U-Boot and Linux need to share the same structure layout and Linux's
spin lock size may vary in the define range, add padding after the U-Boot's
spin lock of the size that accounts for the largest spin lock possible.
This way regardless if Linux is built as RT or not, all the fields (except
the extra part of the spin lock, which is irrelevant from U-Boot's point
of view) will be aligned.

Adjust the padding before the `console_addr` field so it starts at the
desired offset after the change explained above.

Signed-off-by: Wojciech Zmuda <wzmuda@marvell.com>
Change-Id: Ib3fea1ff3e884346215464c8341a80347f9ee06e
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/86825
Reviewed-by: Rick Farrington <rfarrington@marvell.com>
Reviewed-by: Aaron Williams <awilliams@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>