lib: efi_tcg2: Fix SubmitCommand spec violation
Note from EFI Protocol spec:
"The caller is responsible for building the command byte stream
to be sent to the TPM and is also responsible for interpreting
the resulting byte stream returned by the TPM.

Note that the returned status codes reflect the outcome of the
function invocation and not the success (or failure) of the
underlying TPM command."

tpm2_submit_command library call returns status of underlying
TPM command instead of xfer status and is a violation.
Fix it with a function variant to return xfer status as existing
call is widely used in U-Boot command layer.

Change-Id: I61c19177fecdd8175ed4952b6ba19330eb0dd4f2
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/87699
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>