board: cn10k: enable PCIe serial console
Enable serial console over PCIe for CN10ka, CN10kb, CNF10ka
and CNF10kb platforms.

The lock.h file is necessary for implementing exclusive access
to the console memory buffer. The file is identical to its T9x
counterpart. It's not a shared file since it is a low level
piece of code that may require modifications with
per-architecture granularity in future.

CN10k platforms preserve RAM contents on warm reboot and the
console driver's logic does not handle that case very well.
When the console is connected and the board is warm rebooted,
the console connection is broken. It can only be recovered
after cold reboot.

Change-Id: Iab79b27f22e2b37ec89e1f4a80993dafc31c69a0
Signed-off-by: Wojciech Zmuda <wzmuda@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/88890
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>