octeontx: pcie console: zero-initialize the nexus
Nexus is the main struct of the console memory area. The driver
assumes it will be either initialized from the previous boot or
zero-initialized if not yet used. This behavior works well on
T9x but on on CN10k unused memory is initialized to 0x5a5a5a
instead.

Explicitly initialize it to zero if the driver detects it was
not initialized by previous run of the driver. Thanks to that,
subsequent initialization steps can check fields for the value
of 0 and make decision based on that.

Signed-off-by: Wojciech Zmuda <wzmuda@marvell.com>
Change-Id: Ib48b60a18fdb9bc2b9131ecbd4d06b65bb0234e8
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/88888
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>