octeontx: pci-console: fix console operation after U-Boot reset on CN10K
Whenever the CN10K platform was reset from U-Boot, the console driver
would initialize a different console (alternating between 0 & 1).
This was due to the lack of memory initialization.  After a soft-boot
on the CN10K, the console driver initialization would see a valid,
existing console nexus Signature and would skip initialization.
Later, the code would [mis]interpret the existing 'in-use' field
and initialize a different console number.  If console #0 had
previously been in-use, the driver would then initialize console #1.
If console #1 had previously been in-use, the driver would then be
able to initialize console #0 (alternating between consoles 0 & 1).

Fix this problem by checking the 'owner_id' field when an existing
console is detected to be 'in-use'.  If the existing console is
owned by U-Boot, [re]initialize that same console number.

Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Change-Id: Ib6b9730ac8a35967dd1542670fcb894c29af1418
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/92663
Reviewed-by: Aaron Williams <awilliams@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>