octeontx: serial: issue warning if incorrect 'stdin' setting detected
The 'stdin' environment variable must include the pci-bootcmd device
name in order for the host utils to work with it - example:
  stdin=serial,pci-bootcmd@0x027ff000,pci-console@0

If the device name is incorrect (mis-spelled, incorrect address, etc),
the device will NOT be connected to stdin and the pci bootcmd
functionality will not be available to the host.

Ensure the setting is correct during driver initialization by
comparing the 'stdin' setting to the actual device name.  If the
setting is incorrect, issue a warning message.

Signed-off-by: Rick Farrington <rfarrington@marvell.com>
Change-Id: Ib045b60f4b37619b14f854522d9b51d4dedfafe0
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/94801
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Pragnesh Patel <pragneshp@marvell.com>