common/spl: change default u-boot data partition offset for mvebu
Change the default mmc card uboot offset on mvebu platform from 0 to 1
to fix booting from microSD cards.
This was the default before switching spl to the bootrom framework and
other projects (openwrt) relied on it.

Quote from Kconfig:
On some platforms SPL location depends on hardware partition. The ROM
code skips the MBR sector when loading SPL from main hardware data
partition. This adds offset to the main U-Boot image. Set this symbol
to the number of skipped sectors.

Fixes: 2226ca17348 ("arm: mvebu: Load U-Boot proper binary in SPL code based on kwbimage header")

Signed-off-by: Josua Mayer <josua@solid-run.com>