MMC: Added ability to downgrade from HS400 enhanced strobe
In u-boot the call to mmc-deinit would fail if the mmc card
was in HS400ES (enhanced strobe) mode. Code was added that will
disable enhanced strobe mode before downgrading to HighSpeed
mode. This resolves the "device_remove: Device 'sdhci@824000000000.blk'
failed to remove, but children are gone" error message when booting
to the linux kernel.

Signed-off-by: Paul Way <pway@marvell.com>
Change-Id: I5c5cf3f0ffb66b0dc23cc67c253aadbad10bcdd3
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/102370
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>