MMC: Fixed downgrade spped from HS400es to mmc_deinit
In u-boot the call to mmc-deinit would fail if the mmc card
was in HS400ES (enhanced strobe) mode. Code was added that will
disable enhanced strobe mode before downgrading to HighSpeed
mode. This resolves the "device_remove: Device
'sdhci@824000000000.blk'failed to remove, but children are gone"
error message when booting to the linux kernel.

Signed-off-by: Paul Way <pway@marvell.com>
Change-Id: Ie56d05ea5f28afdf51b61f3536718015f3657c1d
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/102424
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>