sr-rz: Change default bus-width for sdhi0
At the u-boot stage of boot we don't know if we are booting
from eMMC or uSDHC. Having the bus-width set to 8 for eMMC
provides more performacne but limits the uSDHC booting to 1-bit
bus width which is slower. A bus width of 4 allows both
boot configuration to operate at a reasonable speed. This will
then be further optimized when booting Linux and devices will
run at their full speed.

Signed-off-by: Jon Nettleton <jon@solid-run.com>