sr-rz: Consolidate env and bootargs handling
This patch consolidates similar bootargs and
environment partitions for all the rzg and rzv
solidrun platforms. We use a custom handler
and choose the partition depending on if we have
booted from usdhc or emmc. If it is usdhc then
we always use partition 0, if it is eMMC we use
the defined ENV partition in the kconfig. I have
moved this to the first boot partition by default
so the second partition is completely available
for the customers use.

Signed-off-by: Jon Nettleton <jon@solid-run.com>