sr-rz: Update the device-tree sdio fixup routine
With recent kernel changes we can now support UHS
speeds for the uSDHC booting as well as eMMC. Add
the properties to support this. As part of this the
vccq pin is converted to a regulator and managed
by the driver, so we only need to update that property
for legacy device trees where the gpio-hog node still
exists.

Signed-off-by: Jon Nettleton <jon@solid-run.com>