Add support to update firmware for eMMC
When updating eMMC/SD devices, ATF requires that the caller initialize
the eMMC/SD host controller and needs to pass a few parameters to ATF
such as the RCA value in order for it to work.

These parameters are passed via the chip select (CS) parameter which is
normally only used for SPI devices.

This change makes the update command update the MMC subsystem and passes
the information on to ATF.

This also removes fields in the header file that are not used, such as old
descriptor formats and function prototypes that do not apply to U-Boot.

Change-Id: Iee55c554b98ca6614f450df8dd0b9e2291c0bef1
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/123950
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Chandrakala Chavva <cchavva@marvell.com>