Enable board checking during updates
Update files now contain information about the board and chip ID
that is supported.  A new flag was introduced into ATF which will
now verify this if the flag is set.  The flag bit was previously
unused so the update protocol was otherwise unchanged.  If ATF
does not support this feature then it will be ignored.

This flag is set by default to prevent incompatible images from
being installed.

Change-Id: I8962805e75ad94c297adfe87fc2b782f9582be67
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/127999
Tested-by: Chandrakala Chavva <cchavva@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>