board: solidrun: solidsense-n8: fix phy detection running twice
Before starting kernel board code attempts to identify the ethernet phy
address for patching Linux DTB.
The fec_get_miibus function allocates structures and can not be called
twice for the same bus. Store the mdio bus handle in a static variable
and reuse it when the phy detection function is called again.

This fixes dtb patching after u-boot failed to start the first boot
target.

Signed-off-by: Josua Mayer <josua@solid-run.com>