serial: sh: add R9A09G056 support to generic init exclusion list
Extend the exclusion list in sh_serial_init_generic() to cover
CONFIG_R9A09G056. This aligns the handling of R9A09G056 with
R9A09G057 and other SoCs that should skip the HSCIF initialization
sequence.

Without this, the driver may attempt to access registers that are
not applicable to R9A09G056, leading to incorrect configuration.

Signed-off-by: John Vincent <john.vincent.xa@bp.renesas.com>